requirements.txt
aiogram==3.19.0
python-dotenv
aiofiles
APScheduler
fastapi
matplotlib
apscheduler
reportlab
